negatively wrapped convolution; updated makefile in examples/users_guide

+++ src/lib/foundations/globals/globals.h
@@ -22,12 +22,12 @@ namespace foundations {
 // function_command.cpp
 // #############################################################################
 
-//! an individual command in function expressed in reverse polish notation
+//! an individual command which is part of a function expressed in reverse polish notation
 
 class function_command {
 public:
 
-	double type;
+	int type;
 	// 1 = push labeled constant
 	// 2 = push immediate constant
 	// 3 = push variable
